03-data-flow/data-publisher/requirements.txt
paho-mqtt==3.1.1
